requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
cookiecutter==2.6.0
loguru==0.7.3
typer==0.16.0
torch==2.7.0
datasets==3.6.0
scikit-learn==1.6.1
matplotlib==3.10.3
transformers==4.52.4
accelerate==1.7.0
google-api-python-client==2.173.0
google-cloud-storage==3.1.1
dvc-gs==3.0.2
fsspec==2025.3.0
dvc==3.60.1
hydra-core==1.3.2
wandb==0.21.0
httpx==0.28.1
locust==2.37.13
onnx==1.18.0
onnxruntime==1.22.1
onnxscript==0.3.2
bentoml==1.4.17
streamlit==1.46.1
evidently==0.7.10
prometheus-client==0.22.1
